Simplify everything
