+++ Sources/PerformanceTesting/Benchmark.swift
@@ -5,6 +5,8 @@
 //  Created by James Bean on 8/11/17.
 //
 
+import Foundation
+
 public struct Benchmark <Subject> {
 
     // MARK: - Type Methods
@@ -18,11 +20,11 @@ public struct Benchmark <Subject> {
     {
         var points: [TestPoint] = []
         for size in testPoints {
-            var trials: [Trial] = []
+            var trials: [Double] = []
             for _ in 0..<trialCount {
                 var subject = setup(size)
                 let time = measure { operation(&subject) }
-                trials.append(TimeTrial(time: time))
+                trials.append(time)
             }
             points.append(TestPoint(size: size, trials: trials))
         }
@@ -36,23 +38,23 @@ public struct Benchmark <Subject> {
         measuring operation: @escaping (Subject) -> Void
     ) -> Benchmark
     {
-        return Benchmark(
-            testPoints: testPoints.map { size in
-                TestPoint.nonMutating(
-                    trialCount: trialCount,
-                    size: size,
-                    setup: setup,
-                    measuring: operation
-                )
+        var points: [TestPoint] = []
+        for size in testPoints {
+            var trials: [Double] = []
+            let subject = setup(size)
+            for _ in 0..<trialCount {
+                trials.append(measure { operation(subject) })
             }
-        )
+            points.append(TestPoint(size: size, trials: trials))
+        }
+        return Benchmark(testPoints: points)
     }
 
     // MARK: - Instance Properties
 
     // temp
     var data: [(Double,Double)] {
-        return testPoints.map { (Double($0.size), $0.average) }
+        return testPoints.map { (Double($0.size), $0.trials.average) }
     }
 
     let testPoints: [TestPoint]
@@ -62,4 +64,12 @@ public struct Benchmark <Subject> {
     init(testPoints: [TestPoint]) {
         self.testPoints = testPoints
     }
+}
+
+/// - Returns: The amount of time that it takes to perform the given `operation`.
+internal func measure (operation: () -> Void) -> Double {
+    let start = CFAbsoluteTimeGetCurrent()
+    operation()
+    let finish = CFAbsoluteTimeGetCurrent()
+    return finish - start
 }
